IOS-2: Добавил задержку для моков
